refactor: Contextクラスだけを観たかったので、インターフェイスの列挙を分離。
